pathing changes + snippet added
